File: Ensure `getRelativePath` returns correct result
Before this commit, there was a string replace of the base path. This
was not ideal as if the file path contained the same as the base path,
this would be completely replaced.

Take the following example with the previous code:

```
$base = "/example"
$fileName = "/example/file/example/test.php"
```

By calling `getRelativePath`, this would incorrectly return
`file/test.php`.